refactor: move SearchContainer from HomePage to its own file component
