+++ frontend/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\App.tsx":"1","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\header\\HeaderStyles.ts":"2","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\index.tsx":"3"},{"size":274,"mtime":1606957775283,"results":"4","hashOfConfig":"5"},{"size":112,"mtime":1606960265832,"results":"6","hashOfConfig":"5"},{"size":196,"mtime":1606960438804,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ca6d0p",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\App.tsx",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\header\\HeaderStyles.ts",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\index.tsx",[]]
+[{"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\App.tsx":"1","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\header\\HeaderStyles.ts":"2","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\index.tsx":"3","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\header\\Header.tsx":"4","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\HomePageStyles.ts":"5","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\HomePage.tsx":"6","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\search_container\\SearchContainerStyles.ts":"7","C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\search_container\\SearchContainer.tsx":"8"},{"size":274,"mtime":1606957775283,"results":"9","hashOfConfig":"10"},{"size":112,"mtime":1606960265832,"results":"11","hashOfConfig":"10"},{"size":196,"mtime":1606960438804,"results":"12","hashOfConfig":"10"},{"size":314,"mtime":1606960695192,"results":"13","hashOfConfig":"10"},{"size":824,"mtime":1606960950255,"results":"14","hashOfConfig":"10"},{"size":1113,"mtime":1606961173249,"results":"15","hashOfConfig":"10"},{"size":1251,"mtime":1606960962666,"results":"16","hashOfConfig":"10"},{"size":998,"mtime":1606961259565,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ca6d0p",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\App.tsx",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\header\\HeaderStyles.ts",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\index.tsx",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\header\\Header.tsx",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\HomePageStyles.ts",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\HomePage.tsx",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\search_container\\SearchContainerStyles.ts",[],"C:\\Users\\tiago\\Desenvolvimento\\EnterpriseChallenges\\front-end-test\\frontend\\src\\pages\\home\\components\\search_container\\SearchContainer.tsx",[]]
